+++ src/components/Node.tsx
@@ -12,6 +12,7 @@ import { NodeContext } from "../context";
 
 type NodeProps = {
   children: React.ReactNode;
+  visible?: Prop<boolean>;
   pos?: Prop<Position>;
   offset?: Prop<Position>;
   rounded?: boolean;
@@ -21,6 +22,8 @@ type NodeProps = {
 
 export const Node: React.FC<NodeProps> = ({ name, children, timeScale, rounded, ...props }) => {
   const parent = useContext(NodeContext);
+  const isVisible = props.visible === undefined ? (parent.visible === undefined ? true : parent.visible) : props.visible;
+  const visible = useProperty<boolean>(isVisible);
   const pos = useProperty<Position>(props.pos || parent.pos || [0, 0]);
   const offsetPos = useOffsetPosition(pos, props.offset || [0, 0]);
   const roundedPos = useIntegerPosition(offsetPos);
@@ -30,6 +33,7 @@ export const Node: React.FC<NodeProps> = ({ name, children, timeScale, rounded, 
     ...node,
     debug: parent.debug,
     pos: rounded ? roundedPos : offsetPos,
+    visible,
     name,
   }), [node, parent.debug, pos]);