+++ src/context/NodeContext.ts
@@ -4,6 +4,7 @@ import { Prop, Position, UpdateFunction, RenderFunction, TickerFunction, UpdateO
 type NodeContextProps = {
   debug?: boolean;
   pos?: Prop<Position>;
+  visible?: Prop<boolean>;
   registerTicker: (id: string, fn: TickerFunction) => void;
   registerUpdate: (id: string, fn: UpdateFunction, options?: UpdateOptions) => void;
   registerRender: (id: string, fn: RenderFunction) => void;
@@ -12,6 +13,7 @@ type NodeContextProps = {
 export const NodeContext = React.createContext<NodeContextProps>({
   debug: false,
   pos: [0, 0],
+  visible: true,
   registerTicker: () => {},
   registerUpdate: () => {},
   registerRender: () => {},